Revert "make a deepcopy of BS document to avoid making changes to original. add tests"
This reverts commit 46981844655f5e6d8dec0bad1273ef3af6985c61.

we still want to switch to deepcopy to avoid making changes to the original BS document, but we haven't yet ironed out outstanding bugs (eg #108), performance regressions, or if/how to let users opt into or out of this behavior.

sooo...we're punting on it for the 1.1.2 release, and we'll revisit after.